src/19/requirements.txt
PyExifTool==0.5.5
PyMySQL==1.0.2
